Resolve #331 -- Disconnected player walks to the fountain (#985)
* When player disconnects he walks to the fountain
* When player reaches the fountain he receives 99999 armor and magic resistance

+++ GameServerCore/IGame.cs
@@ -13,6 +13,6 @@ namespace GameServerCore
 
         bool SetToExit { get; set; }
 
-        bool HandleDisconnect(int userId);
+        bool HandleDisconnect(ulong userId);
     }
 }
